+++ src/Universum/List/Reexport.hs
@@ -10,8 +10,8 @@ module Universum.List.Reexport
 
 import Data.List (break, cycle, drop, dropWhile, filter, genericDrop, genericLength,
                   genericReplicate, genericSplitAt, genericTake, group, inits, intercalate,
-                  intersperse, isPrefixOf, iterate, permutations, repeat, replicate, reverse, scanl,
-                  scanr, sort, sortBy, sortOn, splitAt, subsequences, tails, take, takeWhile,
+                  intersperse, isPrefixOf, iterate, map, permutations, repeat, replicate, reverse,
+                  scanl, scanr, sort, sortBy, sortOn, splitAt, subsequences, tails, take, takeWhile,
                   transpose, unfoldr, unzip, unzip3, zip, zip3, zipWith, (++))
 import Data.List.NonEmpty (NonEmpty (..), head, init, last, nonEmpty, tail)