+++ src/Universum/String/Conversion.hs
@@ -189,7 +189,7 @@ show x = fromString (Show.show x)
 {-# SPECIALIZE show :: Show.Show  a => a -> String  #-}
 
 
- -- | Type class for lazy-strict conversions.
+-- | Type class for lazy-strict conversions.
 class LazyStrict l s | l -> s, s -> l where
     toLazy :: s -> l
     toStrict :: l -> s