+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Change log
   the value.
 * [#18](https://github.com/kowainik/universum/issues/18):
   Add `LazyStrict` type class for conversions.
+* `map` is not `fmap` anymore. Reexport `map` from `Data.List`
 
 
 `universum` uses [PVP Versioning][1].